treewide: use size_t for length variables to avoid implicit narrowing
Several local length variables stored the result of size_t-returning
functions (strlen, blob_pad_len, blob_raw_len, array_list_length) in an
int, which GCC's -Wconversion reports as a potentially value-changing
narrowing conversion.

Widen these variables (and one static helper and one static function
parameter) to size_t where the int type was not needed:

 - avl-cmp: _min() and the memcmp() length in avl_blobcmp()
 - blob: len/delta in blob_fill_pad()
 - blobmsg_json: loop counters in blobmsg_add_array() and the length
   parameter of blobmsg_format_json_list()
 - jshn: loop counters in add_json_array()

udebug_entry_printf() stored the int result of udebug_entry_vprintf() in
a size_t; use int there instead.

No functional change.

Link: https://github.com/openwrt/libubox/pull/46
(cherry picked from commit 73a21977c52a47f9a0ccb6116553083b118d0084)

+++ blob.c
@@ -125,8 +125,8 @@ void
 blob_fill_pad(struct blob_attr *attr)
 {
 	char *buf = (char *) attr;
-	int len = blob_pad_len(attr);
-	int delta = len - blob_raw_len(attr);
+	size_t len = blob_pad_len(attr);
+	size_t delta = len - blob_raw_len(attr);
 
 	if (delta > 0)
 		memset(buf + len - delta, 0, delta);
